* changed StateDatabase::evalInt() to return 0 for NaN values
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@20397 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/bzflag.cxx
@@ -1334,7 +1334,7 @@ int postWindowInit ( void )
 
     SOUNDSYSTEM.startup();
     if (startupInfo.hasConfiguration && BZDB.isSet("volume"))
-      SOUNDSYSTEM.setVolume(static_cast<int>(BZDB.eval("volume"))*0.1f);
+      SOUNDSYSTEM.setVolume(BZDB.eval("volume") * 0.1f);
 
     BZDB.set("soundsystem",SOUNDSYSTEM.name()); // save the sound system they get
   }
